create index query on all objects with search
Create function that filters the parameters based on a search query and
returns the indices of the objects that match the text based search. The
function should be used in the other search queries to reuse one
function for the same functionality.

It is only used for delete mode for now.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>